refactor: replace preview resource fragment with composable
